refactors TriviaGenerator class to its own file
